section 1 updated & section 2 added
