+++ src/App.css
@@ -1,13 +1,13 @@
 .hero {
   height: 80vh;
-  background-image: url("./image/GOW\ PP1.jpeg");
+  /* background-image: url("./components/images/GOW\ PP1.jpeg"); */
   background-size: 100%;
   background-repeat: no-repeat;
 }
 @media (max-width: 768px) {
   .hero {
     background-position: top;
-    height: 80vh;
+    height: 100vh;
     background-size: cover;
   }
 }
