--- src/components/Home/home.css
@@ -1,33 +1,0 @@
-.container {
-  padding: 2em;
-  margin: 0 auto;
-  width: 90%;
-}
-.intro-content {
-  width: 50%;
-}
-
-.btn {
-  display: inline-block;
-  text-decoration: none;
-  color: white;
-  background-color: var(--mainColor);
-  padding: 1em 2em;
-  cursor: pointer;
-}
-
-a:hover {
-  color: var(--secondaryColor);
-}
-
-nav a:hover {
-  color: var(--secondaryColor);
-}
-.hero-title {
-  font-size: 6vmin;
-  text-transform: uppercase;
-}
-.hero-text {
-  font-size: 2.5vh;
-  margin: 1.5rem 0;
-}
